requirements.txt
streamlit
pandas
plotly
bcrypt
python-dotenv
sqlmodel
psycopg2-binary